API Responses are now Events. It is now possible to move between rooms.

+++ src/main/java/fi/beeblebrx/mud/game/GameStateImpl.java
@@ -6,7 +6,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 public class GameStateImpl implements GameState {
-    private ConcurrentMap<Long, Player> players;
+    private ConcurrentMap<Integer, Player> players;
 
     public GameStateImpl() {
         players = new ConcurrentHashMap();
